[COREV] Fix generation for clipu from betweenu pattern
CORE-V uses a definition of signed different from LLVM in that a conversion from
signed to unsigned in COREV implies that a number less than zero will be truncated
to 0, rather than interpreting a negative signed number as a non-converting cast to
a positive unsigned as LLVM does.

This patch reconciles these differences within the definition of the betweenu pattern
so that generation of the clipu instruction canw occur. This fixes a broken test.