[RISCV] Fix generation for clipu from betweenu pattern (#25)
It appears the definition for the betweenu pattern is
erroneous as the use of an unsigned appears to refer to
how the output appears and not how the inputs are treated.
This causes further problems as the tests for clipu expect
this erroneous behaviour.

This patch resolves this issue within the definition of
the betweenu pattern and the tests so that generation of
the clipu instruction can occur and is tested for successfully.
This fixes a failing test.